refactor: move stream chunk assembler
